230512 optimization
