+++ firstSolar/front/src/components/header/Container.jsx
@@ -1,13 +1,36 @@
-import { useEffect } from "react";
+import { useEffect, useRef, useState } from "react";
+import { Link, useLocation, useNavigate } from "react-router-dom";
+import { useDispatch } from "react-redux";
+import { useAccount } from "wagmi";
+
 import HeaderComponent from "./Component";
-import { useDispatch, useSelector } from "react-redux";
 import { setConnect } from "../../modules/connect.js";
 import { setAccount } from "../../modules/account.js";
-import { useNavigate } from "react-router-dom";
 import { setLogin } from "../../modules/login";
 
 const HeaderContainer = () => {
+  const [inputValue, setInputValue] = useState();
+  const [view, setView] = useState(false);
+  const [searchView, setSearchView] = useState(false);
+
   const dispatch = useDispatch();
+  const { pathname } = useLocation();
+  const ref = useRef();
+  const navigate = useNavigate();
+  const { address } = useAccount();
+
+  const handleInputChange = (e) => {
+    const inputValue = e.target.value;
+    const sanitizedValue = inputValue.replace(/[ㄱ-ㅎㅏ-ㅣ가-힣]/g, "");
+    setInputValue(sanitizedValue);
+  };
+
+  const handleSearch = () => {
+    const searchQuery = encodeURIComponent(inputValue);
+    navigate("/searchRedirect", { state: { searchData: searchQuery } });
+
+    setInputValue("");
+  };
 
   useEffect(() => {
     (async () => {
@@ -19,45 +42,39 @@ const HeaderContainer = () => {
         dispatch(setAccount(_account));
         dispatch(setLogin(true));
         dispatch(setConnect(true));
-        // if (document.cookie.split(":")[0] == "metamask") {
-        //   window.ethereum.providers.map(async (item, idx) => {
-        //     if (item.isMetaMask == true) {
-        //       const [_account] = await item.request({
-        //         method: "eth_requestAccounts",
-        //       });
-        //       dispatch(setAccount(_account));
-        //       dispatch(setLogin(true));
-        //       dispatch(setConnect( true ));
-        //     }
-        //   });
-        // } else if (document.cookie.split(":")[0] == "trust") {
-        //   window.ethereum.providers.map(async (item, idx) => {
-        //     if (item.isTrustWallet == true) {
-        //       const [_account] = await item.request({
-        //         method: "eth_requestAccounts",
-        //       });
-        //       dispatch(setAccount(_account));
-        //       dispatch(setLogin(true));
-        //       dispatch(setConnect(true ));
-        //     }
-        //   });
-        // } else if (document.cookie.split(":")[0] == "coinbase") {
-        //   window.ethereum.providers.map(async (item, idx) => {
-        //     if (item.isCoinbaseWallet == true) {
-        //       const [_account] = await item.request({
-        //         method: "eth_requestAccounts",
-        //       });
-        //       dispatch(setAccount(_account));
-        //       dispatch(setLogin(true));
-        //       dispatch(setConnect(true ));
-        //     }
-        //   });
-        // }
       } catch (error) {
         console.error(error);
       }
     })();
   }, []);
-  return <HeaderComponent></HeaderComponent>;
+
+  useEffect(() => {
+    setView(false);
+  }, [pathname]);
+
+  useEffect(() => {
+    const checkIfClickedOutside = (e) => {
+      if (view && ref.current && !ref.current.contains(e.target)) {
+        setView(false);
+      }
+    };
+    document.addEventListener("click", checkIfClickedOutside);
+    return () => {
+      document.removeEventListener("mousedown", checkIfClickedOutside);
+    };
+  }, [view]);
+
+  return (
+    <HeaderComponent
+      handleInputChange={handleInputChange}
+      handleSearch={handleSearch}
+      inputValue={inputValue}
+      navigate={navigate}
+      dispatch={dispatch}
+      address={address}
+      searchView={searchView}
+      setSearchView={setSearchView}
+    ></HeaderComponent>
+  );
 };
 export default HeaderContainer;
