+++ firstSolar/front/src/ui-components/PoolListCom1024px.jsx
@@ -136,7 +136,6 @@ export default function PoolListCom1024px(props) {
   }, [filter]);
 
   React.useEffect(() => {
-    console.log("1024px");
     setMainNetList(Object.keys(mainNet1024px1));
     setPlatformList(Object.values(platform));
   }, []);
